feat: support react-native
React-native has a hobbled URL class that seems to make it's own
WebSocket class cause a fatal crash so stringify it before passing
it in.

Otherwise the URL class is supported everywhere now so no need to
use iso-url any more.

Also noteworthy is that react-native doesn't have any node core
modules so we have to use a similar browser override system to ensure
the pure-js/browser APIs are used instead.