fix: handle null server address
The types for the `Server.address()` method in `@types/ws` are
missing a return type of `null`.  This can happen when a incoming
connection is recieved after the ws `Server` has been closed but
before the underlying HTTP server has finished closing down.

https://github.com/DefinitelyTyped/DefinitelyTyped/pull/70132 will
address the type problem, the change here is to handle the `null`
return value.

It also calls `.destroy` on the incoming message since we'd want
to reclaim the resources immediately and not wait for the remote
to time out.